Add logictest runner
Given a few prerequisites, the script runs the sql logic tests on AWS,
fetches the logs, and sends a summary email.
Logs for now are saved on local disk, with stdout from every instances
attached to the email.

I've added this to my crontab to run nightly and email
daily@cockroachlabs.com