Single-step cluster turnup.
* add a "cockroach-initializer" null resource. Runs only once.
* remove launch.sh in favor of supervisor
* use tiny program that talks to supervisor and gets final status

This is a more reliable way of watching processes and stopping then, as
well as getting the final status (runtime and exit code).

The null_resource setup in terraform can cause issues when instances go
down through anything other than terraform. Still a TODO.
Some of the other logic in the scripts may get moved to go as well.